Add console.log labels to each example.
Hi wife. I just wanted you to see what it looks like when someone sends you a pull request (updates/changes) to one of your repos. This never happens to just stuff you have out there because it is a public repo but when you work on stuff as a team you might need to take in other's changes through a pull request. It could also happen from a rando on the internet if your repo is public and they like to use it to supplement whatever they are working on.